# Cache keys in key-store