This PR adds a LRU cache to key-store.

The cache contains unmarshalled private keys that then get returned quickly from the cache when the key store is queried for a key. This means that the private key stored in the persistent storage as bytes doesn't have to be unmarshalled on every `get()` request thus improving the performance across databases.

In practice, here's a simple, reduced example how the performance improves with this change:

Before:
```
Inserting 1000 documents took 2278 ms, 438 ops/s, 2.278 ms/op
Querying 1000 documents took 81 ms, 12345 ops/s, 0.081 ms/op
```

After:
```
Inserting 1000 documents took 1642 ms, 609 ops/s, 1.642 ms/op
Querying 1000 documents took 72 ms, 13888 ops/s, 0.072 ms/op
```